Iss #14 Track email opens
"Documents by order" page allows to send emails with attached documents.
Module now sends emails with tine image inserted in the body of the
email. This image triggers a function in the email.php which shows that
the email recepient received and opened the email.

Also adds hook to update order status when email was read.